UX: Only have one sidebar open at a time
Previously it supported having open an add/create sidebar and
a details or plaintext editor at the same time.
It didn't support having a plaintext editor and a details pane open at the same time.

This is inconsistent and lead to weird code.
Now, the structure in `Drive.Sidebar.Model` is way clearer
and contains both the add/create and the other sidebars.

This would also be preliminary work for creating an actual sidebar stack.

Closes #98